[Runtime]Move fibers to dedicated header.
