+++ Modules/Luna/Runtime/Coroutine.hpp
@@ -8,7 +8,7 @@
 * @date 2026/2/14
 */
 #pragma once
-#include "Thread.hpp"
+#include "Fiber.hpp"
 
 namespace Luna
 {
