+++ Modules/Luna/Runtime/Thread.hpp
@@ -88,35 +88,6 @@ namespace Luna
     //! @return The thread object of the main thread.
     LUNA_RUNTIME_API IThread* get_main_thread();
 
-    //! @interface IFiber
-    //! Represents a thread executing context that can be switched to 
-    //! for a specified thread.
-    struct IFiber : virtual Interface
-    {
-        luiid("{86a46fbf-43fd-40f3-9aa4-3aa10fa7d6f9}");
-
-    };
-
-    //! Creates a new fiber object.
-    //! @param[in] stack_size The stack size in bytes of the new fiber object.
-    //! @param[in] entry_func The entry function of the fiber.
-    //! @param[in] param The parameter that will be passed to `entry_func` when the fiber is 
-    //! activated.
-    LUNA_RUNTIME_API R<Ref<IFiber>> new_fiber(usize stack_size, void(*entry_func)(void* param), void* param);
-
-    //! Creates a fiber context for current thread.
-    //! @return Returns the fiber context created for the current thread.
-    //! Calling this function multiple times on the same thread does not create multiple fiber contexts, 
-    //! instead the same fiber context will be returned.
-    LUNA_RUNTIME_API R<Ref<IFiber>> convert_thread_to_fiber();
-
-    //! Frees the fiber context for the current thread.
-    LUNA_RUNTIME_API RV convert_fiber_to_thread();
-
-    //! Sets the context of the current thread to the specified fiber.
-    //! @param[in] fiber The fiber to switch to.
-    LUNA_RUNTIME_API void switch_to_fiber(IFiber* fiber);
-
     //! Suspends current thread for a specific period of time. 
     //! @details The actual suspended time may be longer than required.
     //! @param[in] timeMillisecounds The time, in milliseconds, that this thread needs to suspend.
